add compilationdabase functions to core
